+++ src/bittytax/conv/output_csv.py
@@ -91,12 +91,15 @@ class OutputCsv(OutputBase):
         TrType.REFERRAL: "Referral",
         TrType.CASHBACK: "Cashback",
         TrType.FEE_REBATE: "FeeRebate",
+        TrType.MARGIN_GAIN: "MarginGain",
         TrType.WITHDRAWAL: "Withdrawal",
         TrType.SPEND: "Purchase",
         TrType.GIFT_SENT: "Gift",
         TrType.GIFT_SPOUSE: "Spouse",
         TrType.CHARITY_SENT: "Donation",
         TrType.LOST: "Lost",
+        TrType.MARGIN_LOSS: "MarginLoss",
+        TrType.MARGIN_FEE: "LendingFee",
         TrType.TRADE: "Trade",
     }