+++ src/bittytax/transactions.py
@@ -317,6 +317,7 @@ class Buy(TransactionBase):  # pylint: disable=too-many-instance-attributes
         TrType.REFERRAL,
         TrType.CASHBACK,
         TrType.FEE_REBATE,
+        TrType.MARGIN_GAIN,
         TrType.TRADE,
     }
 
@@ -435,6 +436,8 @@ class Sell(TransactionBase):  # pylint: disable=too-many-instance-attributes
         TrType.GIFT_SPOUSE,
         TrType.CHARITY_SENT,
         TrType.LOST,
+        TrType.MARGIN_LOSS,
+        TrType.MARGIN_FEE,
         TrType.TRADE,
     }