+++ CHANGELOG.md
@@ -57,6 +57,10 @@
 - Kraken parser: added "adjustment" type.
 - Conversion tool: added parser for Gate.io exchange. ([#152](https://github.com/BittyTax/BittyTax/issues/152))
 - Crypto.com parser: added "finance.lockup.dpos_compound_interest.crypto_wallet" and finance.lockup.dpos_lock.crypto_wallet" transaction kinds. ([#356](https://github.com/BittyTax/BittyTax/issues/356))
+- KuCoin parser: added "Spot Orders" and "Account History_Funding" files.
+- Kraken parser: added new export format. ([#355](https://github.com/BittyTax/BittyTax/issues/355))
+- Kraken parser: added "dividend" type.
+- Kraken parser: added "adjustment" type.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))